requirements.txt
streamlit==1.40.1
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.0.3
matplotlib==3.7.2
msal==1.28.0
azure-data-tables==12.4.0
APScheduler==3.10.4
